+++ package.json
@@ -26,9 +26,9 @@
   "devDependencies": {
     "@types/jest": "^26.0.0",
     "@types/uuid": "^8.0.0",
-    "@typescript-eslint/eslint-plugin": "^3.2.0",
-    "@typescript-eslint/parser": "^3.2.0",
-    "aws-sdk": "^2.696.0",
+    "@typescript-eslint/eslint-plugin": "^3.3.0",
+    "@typescript-eslint/parser": "^3.3.0",
+    "aws-sdk": "^2.697.0",
     "aws-sdk-mock": "^5.1.0",
     "eslint": "^7.2.0",
     "eslint-config-airbnb-base": "^14.2.0",
