+++ src/constants.ts
@@ -3,3 +3,8 @@
  */
 export const CONNECTIONS_TABLE = 'ConnectionsTable';
 export const GAMES_TABLE = 'GamesTable';
+
+/**
+ * Default Document Version
+ */
+export const DEFAULT_DOCUMENT_VERSION = 1;
