+++ src/functions/onLeaveGame.ts
@@ -1,18 +1,20 @@
 import { Handler } from 'aws-lambda';
-import { removeUserFromGame } from '../module/db';
+import { removeUserFromGame } from '../module/gameDBService';
 import {
-  WebSocketAPIGatewayEvent,
-  LambdaEventBody,
-  LambdaResponse,
-  LambdaEventBodyPayloadOptions,
   GameStates,
+  LambdaEventBody,
+  LambdaEventBodyPayloadOptions,
+  LambdaResponse,
+  WebSocketActions,
+  WebSocketAPIGatewayEvent,
 } from '../types';
 import { response } from '../utils/response';
 import { Logger } from '../utils/Logger';
 import { WebSocketClient } from '../WebSocketClient';
-import { successWebSocketResponse, failedWebSocketResponse, createLeaveResponse } from '../utils/webSocketActions';
-import { broadcastGameUpdate } from '../utils/broadcast';
+import { createLeaveResponse, failedWebSocketResponse, successWebSocketResponse } from '../utils/webSocketActions';
+import { broadcastGameUpdate, broadcastInGameMessage } from '../utils/broadcast';
 import { Game } from '../models/Game';
+import { removeGameDocumentVersion } from '../utils/dbHelper';
 
 /**
  * Handler for joining a game.
@@ -34,7 +36,9 @@ export const handler: Handler = async (event: WebSocketAPIGatewayEvent): Promise
   try {
     // Send success response
     const updatedGame = (await removeUserFromGame(gameId, connectionId)) as Game;
+    removeGameDocumentVersion<Game>(updatedGame);
     console.log('Updated game after leaving a game:', updatedGame);
+
     const res = createLeaveResponse(updatedGame);
     const updatedGameResponse = successWebSocketResponse(res);
     await ws.send(JSON.stringify(updatedGameResponse), connectionId);
@@ -42,9 +46,9 @@ export const handler: Handler = async (event: WebSocketAPIGatewayEvent): Promise
     // Send message to other users in the game
     const { host } = updatedGame;
     if (host) {
-      await broadcastGameUpdate(ws, gameId, GameStates.DELETED);
+      await broadcastGameUpdate(ws, gameId, GameStates.DELETED, connectionId);
     } else {
-      // TODO: Call IN_GAME_UPDATE here
+      await broadcastInGameMessage(ws, gameId, connectionId, WebSocketActions.LEAVE_GAME);
     }
 
     return response(200, 'Joined game successfully');
